requirements.txt
attrs
docopt
pyparsing
python-magic
